feat(front): add module card
