--- frontend/.eslintrc.cjs
@@ -1,35 +1,0 @@
-module.exports = {
-    env: {
-        browser: true,
-        es2021: true,
-    },
-    extends: [
-        'eslint:recommended',
-        'plugin:react/recommended',
-        'plugin:prettier/recommended',
-    ],
-    overrides: [
-        {
-            env: {
-                node: true,
-            },
-            files: ['.eslintrc.{js,cjs}'],
-            parserOptions: {
-                sourceType: 'script',
-            },
-        },
-    ],
-    parserOptions: {
-        ecmaVersion: 'latest',
-        sourceType: 'module',
-    },
-    plugins: ['react'],
-    rules: {
-        'react/prop-types': 'off',
-        'no-var': 'error',
-        'prefer-const': 'warn',
-        'no-unused-vars': 'warn',
-        'react/react-in-jsx-scope': 'off',
-        'prettier/prettier': ['warn', { endOfLine: 'auto' }],
-    },
-};
