--- frontend/.prettierrc
@@ -1,11 +1,0 @@
-{
-    "semi": true,
-    "singleQuote": true,
-    "jsxSingleQuote": false,
-    "bracketSpacing": true,
-    "trailingComma": "all",
-    "printWidth": 80,
-    "tabWidth": 4,
-    "arrowParens": "avoid",
-    "endOfLine": "lf"
-}
