+++ frontend/prettier.config.js
@@ -1,0 +1,3 @@
+export default {
+  singleQuote: true,
+};
