+++ frontend/src/components/App/App.css
@@ -2,7 +2,7 @@ body,
 html {
   padding: 0;
   margin: 0;
-  font-family: 'YS Display', Arial, Helvetica, sans-serif;
+  font-family: 'YS Text', Arial, Helvetica, sans-serif;
 }
 
 #root {
