+++ frontend/src/components/DiagramIllustration/Circle/Circle.jsx
@@ -1,0 +1,7 @@
+import './Circle.sass';
+
+function Circle({ radius, children }) {
+  return <div className={`circle circle_radius_${radius}`}>{children}</div>;
+}
+
+export { Circle };
