+++ frontend/src/components/DiagramIllustration/Circle/Circle.sass
