+++ frontend/src/components/DiagramIllustration/Peace/Peace.sass
@@ -1,8 +1,5 @@
-
 .peace
-
     position: absolute
-    &_active
     &_radius
         &_first
             z-index: 1
@@ -161,8 +158,6 @@
 
         &_blue
             right: 0
-            &.peace_level_junior.peace_radius
-
             &.peace_level_middle.peace_radius
                 &_first
                     bottom: 70px
