+++ frontend/src/components/DiagramList/DiagramList.jsx
@@ -1,7 +1,7 @@
 import './DiagramList.sass';
 
 function DiagramList({ children }) {
-    return <ul className="diagram-list">{children}</ul>;
+  return <ul className="diagram-list">{children}</ul>;
 }
 
 export { DiagramList };
