+++ frontend/src/components/ModuleCard/ModuleCard.sass
@@ -1,0 +1,59 @@
+.module-card
+  display: flex
+  flex-direction: column
+  width: 286px
+  height: 350px
+  padding: 24px 16px
+  border-radius: 16px
+  background: #fff
+  box-shadow: 0 0 20px -2px rgba(56 59 128 / 25)
+  box-sizing: border-box
+  &__top-wrapper
+    display: flex
+    column-gap: 8px
+    margin-bottom: 8px
+    font-size: 11px
+    font-weight: 500
+    line-height: 1.09
+  &__separator
+    width: 4px
+    height: 4px
+    border-radius: 2px
+    background-color: #1A1B22
+    align-self: center
+  &__occupation
+    margin: 0
+  &__type
+    margin: 0
+  &__level
+    margin: 0
+  &__title
+    max-height: 50px
+    height: 100%
+    margin: 0
+    font-size: 16px
+    font-weight: 500
+    line-height: 1.25
+  &__image
+    width: 100%
+    height: 166px
+    margin-bottom: 30px
+    object-fit: contain
+    //object-position: center
+  &__price-wrapper
+    display: flex
+    flex-direction: column
+    align-items: flex-start
+    row-gap: 4px
+  &__price
+    margin: 0
+    font-size: 16px
+    font-weight: 500
+    line-height: 1.25
+    &_type_no-monthly-price
+      margin-top: 43px
+  &__full-price
+    margin: 0
+    font-size: 11px
+    font-weight: 500
+    line-height: 1.09
