+++ frontend/src/pages/LearningTrack/LearningTrack.jsx
@@ -1,14 +1,15 @@
 import { Header } from '../../components/Header/Header';
 import { SidePanel } from '../../components/SidePanel/SidePanel';
 import { Profile } from '../Profile/Profile';
-function LearningTrack(props) {
-    return (
-        <>
-            <Header />
-            <SidePanel />
-            <Profile level="middle" />
-        </>
-    );
+
+function LearningTrack() {
+  return (
+    <>
+      <Header />
+      <SidePanel />
+      <Profile level="middle" />
+    </>
+  );
 }
 
 export { LearningTrack };
